pyproject.toml
[project]
name = "projet-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
